+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 <!-- There is always Unreleased section on the top. Subsections (Add, Changed, Fix, Removed) should be Add as needed. -->
 ## Unreleased
+- Add `Box` module with functions
+    - `createFromStrings`
+    - `concat`
 
 ## 1.0.0 - 2019-01-31
 - Initial implementation